refactor(keystone): reduce keystoneauth usage complexity
This patch set refactors and fixes the use of the keystoneauth to load
the session from cfg.CONF. This removes the unnecessary wrapping of
utility function into a class, but also allows the code to use other
accepted plugins to form the keystoneauth object.

The need to read environment variables should be handled only in the CLI
and should be accounted for by the CLI framework, not in a server utility
function.

Change-Id: Ib086f103bbb1e27fe8228ccf5f0d40526796e1e5
Signed-off-by: Tin Lam <tin@irrational.io>