style(imports): update python style
This patch set imports the constant module rather than the individual
constants. This allows adding future constant into the module easier with
needing to update import, also it is easier for developer to know the
constants are coming from a different module rather than local. Lastly,
this patch set also fixes the alphabetical order of imports.

Change-Id: I886b5368230ab2e11b9792ac7e830ded0fc91d06
Signed-off-by: Tin Lam <tin@irrational.io>