Ensure SSH git clone doesn't prompt for user
This patchset adds additional options to the git SSH command
to ensure that no prompt is required to perform the  clone.
This is needed for automation purposes. (For background when
first trying to clone something via SSH the client asks the
user whether they want to add the host to the hosts files --
this skips that check).

Change-Id: I752a354c5616f40eb8a0dbcb8606370654cabe14