api: Replace conflicting environment variable
The Armada API relies on an environment variable, `ARMADA_API_PORT`,
to start uWSGI with the desired port; however, it conflicts with an
environment variable already in use. This causes the `entrypoint.sh`
script to use the wrong port value for uWSGI when using user-specified
`armada_api` values through Helm (i.e. the port is already set and the
`entrypoint.sh` script does not fallback on its default value). This
commit adds missing uWSGI configuration values to the Armada chart
values and changes their names to avoid potential conflicts.

Change-Id: I00ae3431e36593956705dcdb887025d688d804df