Fail wait when no resources found
This removees a fail-safe that allowed releases which did not
contain pods (intentionally) to still succeed after a best effort
to wait for them until timeout.

Now that we have the ability to disable waiting on resource types
`wait.resources` [0], this fail-safe is no longer needed.

Now when resources are not found, armada will fail with a message
for the user to check their `wait.resources` and labels and
configure as needed. This way we can prompt the user to remove
unnecessary waiting from their deployments.

There is also a longer term plan to make these configurations less
often needed [1].

[0]: https://review.openstack.org/#/c/603901/
[1]: https://review.openstack.org/#/c/636440/

Change-Id: I859326470ecba49f2301705409c51312a601e653