tools: Update Helm to v2.13.1
Helm v2.13.1 has been released [0], and is the next version of Helm
Armada is compatible with. Currently, Armada is not compatible with the
latest version of Helm toolkit due to a divergence caused in Helm v2.13.
This change uplifts Helm to v2.13.1 to restore compatibility with the
latest version of Helm toolkit.

[0] https://github.com/helm/helm/releases/tag/v2.13.1

Change-Id: Ieaf2475562c56530b6ec69c6a43611b4b47b7c83