Fix bandit errors
Bandit 1.6.0 changes the behavior of the '-x' option so that it now
supports glob patterns. Update our tox file to use to correctly exclude
test code from bandit scans. This requires bumping our minimum bandit version.

As an aside, changing the behavior of an option like this in a minor
version if bad form :(

Change-Id: I11bee26373ea9894b47d7f964eac88ef70859732