fix: improve error message when user passes in only api and url
Also change enable to enabled.

Before if a user passed in --api and --url TEXT the error message would
return a confusing message that the user must pass in url when api is
enabled.

Now the error message says the user must provide url and token when api
is enabled.

Change-Id: I3284a7e1f920217dc476abcdd1fc17701c85655e