images: Create single metrics dir in entrypoint
The entrypoint script for the Armada Docker container attempts to create
a nested, temporary directory when one is not provided through an
environment variable. This is fine when deploying Armada via a Helm
chart, as a writable volume mount exists; however, the directory
/tmp/armada/metrics does not exist when running as a standalone
container. This commit changes the entrypoint script to use a flat,
temporary directory to avoid requiring a user to mount a temporary
volume.

Change-Id: I26857908fa90c64c98038d508263a5094b06668a
Signed-off-by: Drew Walters <andrew.walters@att.com>