Fix: Use Foreground deletion
Armada intends to use `propagationPolicy: Foreground` when deleting
resources. However, the empty V1DeleteOptions object in the body of the
delete API call takes priority over the propogation_policy specified as
a query param, resulting in the per-resouce default. For Job and CronJob
resources, the default is Orphan, and for others it is Background.

This change includes the desired propogation_policy in V1DeleteOptions.

Reference: https://kubernetes.io/docs/concepts/workloads/controllers/garbage-collection/#setting-the-cascading-deletion-policy

Change-Id: Iffee12b426ba1e7741eb5bd687ca1b2c11cb071d