[chart_deploy] Handle unknow chart status in a more resilient manner
In cases when chart is in a pending state and last deployment age
for chart is less than the chart wait timeout, armada can consume
all the retries without attempting to purge and re-install the
chart.

This PS addresses the above issue by waiting for the chart until
the wait timeout is met and then proceeding ahead with usual
armada process of purging and re-installing the chart if the status
is not DEPLOYED.

Other code changes in this PS is just re-arranging the existing logic to
fix some edge cases and solve pep8 issues.

Change-Id: Id11d0fb06201264015b3064a43e83903f99a30b1