Fix dependency symlink idempotency
Previously, existing dependency symlinks were causing
FileExistsErrors. This simplifies the logic to ensure
they do get deleted before being recreated. It also
moves to erroring if a chart's `charts` directory is
not a directory, since this is something a chart author
ought to fix, since it needs to be a directory in order
to add dependency sub-directories to it.

Change-Id: I7cc1d50f2bf09602866ce2cf13a4d6c8f6e118ce