Restored ubuntu_bionic image build
This PS restores image build for ubuntu_bionic and adds appropriate
gates to keep it tested by appropriate functional and integratinal
tests.

Also the latest osh-infra commit was used with kubernetes 1.27.1 or
integration tests.

Make images section of Makefile was enriched with run_armada step in
order to quicktest the freshly built images - this is vital for focal
and bionic images differences.

Change-Id: Ida43eebd4aa51fff6261ac9cf5f8259359bc99a9